c0cp: fix default unit size setting
Currently, the default unit size is set to 1MB which means
that all writes with buffer sizes less than 1MB * units in
parity group will do RMWs (read-modify-writes) and the
resulting performance will be degraded.

Solution: 1) set default unit size to 4K. 2) Add -u option
to allow user to specify the optimal unit size.

Change the format of block size from bytes to KiBs also
to align with the format of unit size and for convenience.