c0appz: fix arbitrary-file-size i/o regression
Since commit 9d38ae4 arbitrary-file-size i/o is broken.
In commit cf1fe97 it was fixed for 4K-sizes, here we fix it
for any file-sizes.

The fix is to not change the segment's size when we read
the data from file. The segment size must always be the
multiple of 4K.