c0appz: introduce -a option for optimal m0bs estimation
Currently, m0bs is always estimated which gives the wrong
results for the composite objects.

Solution: introduce -a option for m0bs estimation and use
the specified by user block size for m0bs by default.

In future, we may improve c0appz_m0bs() to make it aware
of the composite objects.