c0appz: fix several local endpoints and proc fids
In commit 9f3a3aa we introduced parameters names in the
startup config file, but broke the functionality when several
local endpoints and proc fids could be specified in it and
the needed one could be selected via the idx argument at
c0appz_init(idx).

Solution: 1) update local endpoint and proc fid parameters
names format in the config file to LOCAL_ENDPOINT_ADDR%d and
LOCAL_PROC_FID%d (where %d is the index starting from 0);
2) fix c0appz_init() to honour the idx parameter;
3) update sage-user-application-assignment script and add
to this repository for the convenience.